requirements.txt
openai
mcp
python-dotenv
requests
